requirements.txt
streamlit==1.37.1
fireworks-ai==0.15.0
